fix: Improve error handling in LineReader.openInEditor (#4)
While the initial commit of the openInEditor function _had_ error
handling, it was insufficient owing to the screen being redrawn. This
commit improves that error handling by ensuring that they will be shown
to the user, and a new prompt being drawn on a new line not overwriting
the error.