fix: Improve robustness of getCursorPos
A transient and difficult to reproduce error would sometimes occur
during the `getCursorPos` function, resulting in only a partial Device
Status Report response being read from stdin.

This commit replaces the implementation with a buffered i/o reader, and
by using the ReadString method, we see to ensure that the full response
is read.